requirements.txt
semver~=2.13.0
PyYAML~=6.0